+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_get_oss_upload_policy_response_body.py
@@ -1,0 +1,204 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from typing import List, Any
+
+from alibabacloud_websitebuild20250429 import models as main_models
+from darabonba.model import DaraModel
+
+class GetOssUploadPolicyResponseBody(DaraModel):
+    def __init__(
+        self,
+        access_denied_detail: str = None,
+        allow_retry: bool = None,
+        app_name: str = None,
+        dynamic_code: str = None,
+        dynamic_message: str = None,
+        error_args: List[Any] = None,
+        module: main_models.GetOssUploadPolicyResponseBodyModule = None,
+        request_id: str = None,
+        root_error_code: str = None,
+        root_error_msg: str = None,
+        synchro: bool = None,
+    ):
+        self.access_denied_detail = access_denied_detail
+        self.allow_retry = allow_retry
+        self.app_name = app_name
+        self.dynamic_code = dynamic_code
+        self.dynamic_message = dynamic_message
+        self.error_args = error_args
+        self.module = module
+        # Id of the request
+        self.request_id = request_id
+        self.root_error_code = root_error_code
+        self.root_error_msg = root_error_msg
+        self.synchro = synchro
+
+    def validate(self):
+        if self.module:
+            self.module.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.access_denied_detail is not None:
+            result['AccessDeniedDetail'] = self.access_denied_detail
+
+        if self.allow_retry is not None:
+            result['AllowRetry'] = self.allow_retry
+
+        if self.app_name is not None:
+            result['AppName'] = self.app_name
+
+        if self.dynamic_code is not None:
+            result['DynamicCode'] = self.dynamic_code
+
+        if self.dynamic_message is not None:
+            result['DynamicMessage'] = self.dynamic_message
+
+        if self.error_args is not None:
+            result['ErrorArgs'] = self.error_args
+
+        if self.module is not None:
+            result['Module'] = self.module.to_map()
+
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+
+        if self.root_error_code is not None:
+            result['RootErrorCode'] = self.root_error_code
+
+        if self.root_error_msg is not None:
+            result['RootErrorMsg'] = self.root_error_msg
+
+        if self.synchro is not None:
+            result['Synchro'] = self.synchro
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AccessDeniedDetail') is not None:
+            self.access_denied_detail = m.get('AccessDeniedDetail')
+
+        if m.get('AllowRetry') is not None:
+            self.allow_retry = m.get('AllowRetry')
+
+        if m.get('AppName') is not None:
+            self.app_name = m.get('AppName')
+
+        if m.get('DynamicCode') is not None:
+            self.dynamic_code = m.get('DynamicCode')
+
+        if m.get('DynamicMessage') is not None:
+            self.dynamic_message = m.get('DynamicMessage')
+
+        if m.get('ErrorArgs') is not None:
+            self.error_args = m.get('ErrorArgs')
+
+        if m.get('Module') is not None:
+            temp_model = main_models.GetOssUploadPolicyResponseBodyModule()
+            self.module = temp_model.from_map(m.get('Module'))
+
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+
+        if m.get('RootErrorCode') is not None:
+            self.root_error_code = m.get('RootErrorCode')
+
+        if m.get('RootErrorMsg') is not None:
+            self.root_error_msg = m.get('RootErrorMsg')
+
+        if m.get('Synchro') is not None:
+            self.synchro = m.get('Synchro')
+
+        return self
+
+class GetOssUploadPolicyResponseBodyModule(DaraModel):
+    def __init__(
+        self,
+        dir: str = None,
+        host: str = None,
+        policy: str = None,
+        security_token: str = None,
+        signature: str = None,
+        version: str = None,
+        xoss_credential: str = None,
+        xoss_date: str = None,
+    ):
+        self.dir = dir
+        self.host = host
+        self.policy = policy
+        self.security_token = security_token
+        self.signature = signature
+        self.version = version
+        # x-oss-credential
+        self.xoss_credential = xoss_credential
+        # x-oss-date
+        self.xoss_date = xoss_date
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.dir is not None:
+            result['Dir'] = self.dir
+
+        if self.host is not None:
+            result['Host'] = self.host
+
+        if self.policy is not None:
+            result['Policy'] = self.policy
+
+        if self.security_token is not None:
+            result['SecurityToken'] = self.security_token
+
+        if self.signature is not None:
+            result['Signature'] = self.signature
+
+        if self.version is not None:
+            result['Version'] = self.version
+
+        if self.xoss_credential is not None:
+            result['XossCredential'] = self.xoss_credential
+
+        if self.xoss_date is not None:
+            result['XossDate'] = self.xoss_date
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Dir') is not None:
+            self.dir = m.get('Dir')
+
+        if m.get('Host') is not None:
+            self.host = m.get('Host')
+
+        if m.get('Policy') is not None:
+            self.policy = m.get('Policy')
+
+        if m.get('SecurityToken') is not None:
+            self.security_token = m.get('SecurityToken')
+
+        if m.get('Signature') is not None:
+            self.signature = m.get('Signature')
+
+        if m.get('Version') is not None:
+            self.version = m.get('Version')
+
+        if m.get('XossCredential') is not None:
+            self.xoss_credential = m.get('XossCredential')
+
+        if m.get('XossDate') is not None:
+            self.xoss_date = m.get('XossDate')
+
+        return self
+