+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_get_temp_download_url_request.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class GetTempDownloadUrlRequest(DaraModel):
+    def __init__(
+        self,
+        oss_key: str = None,
+    ):
+        self.oss_key = oss_key
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.oss_key is not None:
+            result['OssKey'] = self.oss_key
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('OssKey') is not None:
+            self.oss_key = m.get('OssKey')
+
+        return self
+